+++ crates/consensus/primary/src/consensus/tests/consensus_utils.rs
@@ -5,7 +5,7 @@
 use std::sync::Arc;
 use tn_storage::{traits::Database, CertificateStore, ConsensusStore};
 
-pub const NUM_SUB_DAGS_PER_SCHEDULE: u64 = 100;
+pub const NUM_SUB_DAGS_PER_SCHEDULE: u32 = 100;
 
 pub fn make_consensus_store<DB: Database>(db: DB) -> Arc<ConsensusStore<DB>> {
     Arc::new(ConsensusStore::new(db))
